Moved rest methods to a static helper class
